Now works on Daffodil as well as IBM DFDL.
Reorg files using Daffodil standard project layout.

This progress is with a 2.4.0-SNAPSHOT sandbox with changes for
separated sequences and separator suppression. It will be a supported
schema for Daffodil once release 2.4.0 of Daffodil is officially
released.

Note use of tunable emptyElementPolicy emptySuppressed which
provides IBM compatibility for the way empty elements are handled.
(Name of this new property and values of it are subject to change.)

Added config to tdml to suppress noise warnings.

Note that all tests pass using ibmDFDLCrossTester both under eclipse
and via sbt test when the build.sbt is configured to use the cross tester.

Changed to defaultRoundTrip='twoPass' to test unparsing also.

On Daffodil all tests continue to work.

On IBM DFDL Cross tester all tests work for parsing. One test does not
work for unparsing, the test testSA1. According to IBM, regression
test for this schema tests the TLogACE sub-schema for both
parse/unparse, but the TLogSA subschema only for parsing.